# LineReader: Reuse temporary buffer to reduce per-line allocation